+++ Shrike/ElementMatch.cs
@@ -31,9 +31,22 @@ public readonly struct ElementMatch<TElement>
     /// <summary>
     /// A list of delegates which will be called whenever this matcher matches any method during the use of the <see cref="ISequenceMatcher{TElement}.Find(SequenceBlockMatcherFindOccurence, SequenceMatcherRelativeBounds, IReadOnlyList{ElementMatch{TElement}})"/> method.
     /// </summary>
-    public IReadOnlyList<ElementMatchDelegate<TElement>> Delegates { get; init; }
+    public IReadOnlyList<ElementMatchDelegate<TElement>> Delegates { get; init; } = new List<ElementMatchDelegate<TElement>>();
 
-    private Func<TElement, bool> Closure { get; init; }
+    /// <summary>
+    /// The function that tests whether a given element matches this match.
+    /// </summary>
+    public Func<TElement, bool> Closure { get; init; }
+
+    /// <summary>
+    /// A list of delegates that get called each time a find operation is started at a new position of a sequence.
+    /// </summary>
+    public IReadOnlyList<Action>? SetupDelegates { get; init; }
+
+    /// <summary>
+    /// A list of postcondition delegates which decide whether the match was actually successful as a whole.
+    /// </summary>
+    public IReadOnlyList<Func<TElement, bool>>? Postconditions { get; init; }
 
     /// <summary>
     /// Initializes a new instance of the <see cref="ElementMatch{TElement}"/> class.
@@ -44,7 +57,6 @@ public readonly struct ElementMatch<TElement>
     {
         this.Description = description;
         this.Closure = closure;
-        this.Delegates = new List<ElementMatchDelegate<TElement>>();
     }
 
     /// <summary>
@@ -65,8 +77,24 @@ public readonly struct ElementMatch<TElement>
     /// <summary>
     /// Creates a copy of the match with an additional match delegate, which will be called whenever this matcher matches any method during the use of the <see cref="ISequenceMatcher{TElement}.Find(SequenceBlockMatcherFindOccurence, SequenceMatcherRelativeBounds, IReadOnlyList{ElementMatch{TElement}})"/> method.
     /// </summary>
-    /// <param name="delegate"></param>
+    /// <param name="delegate">The delegate.</param>
     /// <returns>A new element match with an additional match delegate.</returns>
     public ElementMatch<TElement> WithDelegate(ElementMatchDelegate<TElement> @delegate)
         => this with { Delegates = this.Delegates.Append(@delegate).ToList() };
+
+    /// <summary>
+    /// Creates a copy of the match with an additional setup delegate, which will be called each time a find operation is started at a new position of a sequence.
+    /// </summary>
+    /// <param name="delegate">The delegate.</param>
+    /// <returns>A new element match with an additional match delegate.</returns>
+    public ElementMatch<TElement> WithSetupDelegate(Action @delegate)
+        => this with { SetupDelegates = this.SetupDelegates is null ? new List<Action> { @delegate } : this.SetupDelegates.Append(@delegate).ToList() };
+
+    /// <summary>
+    /// Creates a copy of the match with an additional postcondition that decides whether the match was actually successful as a whole.
+    /// </summary>
+    /// <param name="postcondition">The postcondition delegate.</param>
+    /// <returns>A new element match with an additional postcondition.</returns>
+    public ElementMatch<TElement> WithPostcondition(Func<TElement, bool> postcondition)
+        => this with { Postconditions = this.Postconditions is null ? new List<Func<TElement, bool>> { postcondition } : this.Postconditions.Append(postcondition).ToList() };
 }